few modifs to make usermanage work & cool

+++ src/Controller/AdminUserController.php
@@ -67,7 +67,10 @@ class AdminUserController extends AbstractController
 
         //    Get user with id
         $user = $em->getRepository(User::class)->find($id);
-
+        // if(in_array('ROLE_ADMIN', $user->getRoles())){
+        //     $this->addFlash('warning', 'Êtes-vous certain de vouloir modifier cet user ?');
+        //     return $this->redirectToRoute('userList');
+        // }
         $form = $this->createForm(UserFormType::class, $user);
         $form->handleRequest($request);
 
@@ -98,15 +101,20 @@ class AdminUserController extends AbstractController
     public function deleteUser(EntityManagerInterface $em, $id)
     {
          // REstrictions admin
-         $this->denyAccessUnlessGranted('ROLE_ADMIN');
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
         $user = $em->getRepository(User::class)->find($id);
-
-        $em->remove($user);
-        $em->flush();
-        $this->addFlash('success', 'Le user a bien été supprimé ;)');
-
-
-        return $this->redirectToRoute('userList');
+       
+        if(in_array('ROLE_ADMIN', $user->getRoles())){
+            $this->addFlash('error', 'Impossible de supprimer ce user');
+            return $this->redirectToRoute('userList');
+        }
+      
+            $em->remove($user);
+            $em->flush();
+            $this->addFlash('success', 'Le user a bien été supprimé ;)');
+    
+    
+            return $this->redirectToRoute('userList');
     }
 
 }
